refactor: rename Notifications screen to InAppNotifications
