+++ src/screens/HomeFlow/InAppNotifications/components/NotificationsList/index.tsx
@@ -7,8 +7,8 @@ import {commonStyles, windowWidth} from '@constants/styles';
 import {useScrollShadow} from '@hooks/useScrollShadow';
 import {Header} from '@navigation/components/Header';
 import {ClearButton} from '@navigation/components/Header/components/ClearButton';
-import {DefaultNotification} from '@screens/Notifications/components/DefaultNotification';
-import {EmptyNotifications} from '@screens/Notifications/components/EmptyNotifications';
+import {DefaultNotification} from '@screens/HomeFlow/InAppNotifications/components/DefaultNotification';
+import {EmptyNotifications} from '@screens/HomeFlow/InAppNotifications/components/EmptyNotifications';
 import {InAppNotificationActions} from '@store/modules/InAppNotifications/actions';
 import {LinkingActions} from '@store/modules/Linking/actions';
 import {ClearIcon} from '@svg/ClearIcon';
